Change location of URI encoder
