+++ pages/dao/[symbol]/proposal/new.tsx
@@ -52,7 +52,7 @@ import { Transaction } from '@solana/web3.js'
 import Loader from '@components/Loader'
 import { TOKR_DAO } from '@components/instructions/tools'
 import useRouterHistory from '@hooks/useRouterHistory'
-import { constructUri } from '@hooks/useUri'
+import { constructUri } from '@utils/resolveUri'
 
 const schema = yup.object().shape({
 	title: yup.string().required('Title is required'),
@@ -413,26 +413,25 @@ const New = (props) => {
 			setLiteMode(false)
 
 			const url = constructUri(router.query?.uri.toString(), true);
-			console.log("\n\n\n\n\n\n\n\n\n\n\n\n\nurl", url)
 
 			fetch(url, {
 				method: 'GET',
 			})
-				.then((res) => res.json())
-				.then((res) => {
-					const temp = res.attributes.map((attribute) => {
-						res[attribute.trait_type] = attribute.value
-					})
-
-					setPropertyData({
-						...res,
-					})
-
-					return res
+			.then((res) => res.json())
+			.then((res) => {
+				const temp = res.attributes.map((attribute) => {
+					res[attribute.trait_type] = attribute.value
 				})
-				.catch((error) => {
-					console.log('error', error)
+
+				setPropertyData({
+					...res,
 				})
+
+				return res
+			})
+			.catch((error) => {
+				console.log('error', error)
+			})
 		}
 	}, [router])
